fix(Map): enable display of elevation textures on iOS (#445)
The root cause was that the combination of linear filtering and floating
point textures is unsupported on iOS because it requires the
`OES_texture_float_linear` extension, which is unsupported.

In that case, we fallback to nearest neighbour filtering, at the cost of
reduced visual quality for elevation related features (terrain,
hillshading, colormaps...).

fix #445