perf(TileMesh): update heightmap only when actually necessary
Updating the heightmap is a costly operation that requires a texture
readback. However, we only need the heightmap when performing a raycast
on the tile, or when querying elevations. Two features that are not used
very frequently in a normal scenario.

Now, the heightmap is only updated when actually necessary: just before
elevation queries, and just before performing a raycast, if the bounding
box of the tile actually intersects the ray.