fix(TileMesh): ensure that min/max coming from heightmap is consistent
If the heightmap is completely flat (e.g when no texture is available
yet), we would compute a min/max from this flat heightmap, giving values
of 0/0. This would render the tile's bounding box completely flat, even
if it initially had a non zero min/max set.

Now we use the min/max computed from the heightmap only if the values
are already within the range defined by the existing min/max.