feat: rewrite DrawTool
The DrawTool class is entirely rewritten. The main difference is that
this tool no longer produces GeoJSON objects, but instances of the Shape
entity.

The Shape entity can be exported to GeoJSON or to OpenLayers features.

BREAKING CHANGE: the `DrawTool` API has changed. It now produces `Shape`
instances. The `Drawing` class is removed in favor of the `Shape` entity.

fix #115
fix #378