fix(Layer): ensure that layers are always registered in the material (#462)
In some cases, it was possible that a layer assigns a texture to a map
tile when this layer was not previously registered to this map tile.

This can happen when the source is dynamically updated and its extent
can change over time, such as `VectorSource` with added and removed
features.

The extent is computed from the existing features, so that some map tiles
might not intersect this extent, and the layer is not registered on them.
However, later a feature might be added, giving a new extent that might
now intersect with those empty map tiles.

fix #462