fix(Tiles3D): fix bounding box display when we have a mesh transform
When activating bboxes (in the inspector for instance), the code searches for node that have the
.userdata.metadata property, then generates a bounding box according to these metadata, and add it
to the current node.

The issue is that this metadata object wasn't added to the Object3D that represents the tile, but
the one that represents the content. If this content has a transform property (which can happen),
the bounding box was getting this transform as well, while it shouldn't.

The solution is to add the medata to the node representing the tile.

NOTE: it's very handy for converters to add a y-up to z-up transform to the gltf (the content),
so that it doesn't need to translate coordinates at conversion time if the source file has
z-up coordinates. That's what py3dtiles does with ifc files for instance.